Improving comments and multiple line strings
We are improving comments and multiple line strings that were
intended to guide the reviewer to see corrections on the code.
There were a lot of comments explaining all the corrections that the
reviewer pointed to, and now, they will be irrelevant to other people
who will read the code. Better to remove them.